lampochkaModern/requirements.txt
flask==2.1.1
paho-mqtt==1.6.1
flask-sqlalchemy==2.5.1

